Improve handling of generic properties.
