+++ spec/support/entities/generic_properties.rb
@@ -1,0 +1,48 @@
+# frozen_string_literal: true
+
+require 'support/entities'
+
+module Spec::Support::Entities
+  module GenericProperties
+    def initialize(**properties)
+      @properties = {
+        'amplitude' => nil,
+        'frequency' => nil
+      }
+
+      super(**properties)
+    end
+
+    attr_reader :properties
+
+    private
+
+    def get_property(key)
+      @properties.fetch(key.to_s) { super(key) }
+    end
+
+    def inspectable_properties
+      @properties
+    end
+
+    def set_properties(properties, force:)
+      matching, non_matching = bisect_properties(properties, self.properties)
+
+      super(non_matching, force: force)
+
+      defaults = {
+        'amplitude' => nil,
+        'frequency' => nil
+      }
+      values = force ? defaults : self.properties
+
+      @properties = values.merge(matching)
+    end
+
+    def set_property(key, value)
+      super(key, value) unless properties.key?(key.to_s)
+
+      properties[key.to_s] = value
+    end
+  end
+end
